Fixed CMake for UWP, added classes/stubs
First commit by Lukas 🤘🏻
