+++ CMakeLists.txt
@@ -63,6 +63,8 @@ set(XWIN_PROTOCOL_PATH "Noop")
 
 if(XWIN_PROTOCOL STREQUAL "WIN32")
     set(XWIN_PROTOCOL_PATH "Win32")
+elseif(XWIN_PROTOCOL STREQUAL "UWP")
+    set(XWIN_PROTOCOL_PATH "UWP")
 elseif(XWIN_PROTOCOL STREQUAL "COCOA")
     set(XWIN_PROTOCOL_PATH "Cocoa")
 elseif(XWIN_PROTOCOL STREQUAL "XCB")
