+++ src/CrossWindow/Common/EventQueue.h
@@ -27,29 +27,3 @@
 #elif XWIN_NOOP
 #include "../Noop/NoopEventQueue.h"
 #endif
-
-namespace xwin
-{
-    class Window;
-
-    class EventQueue
-    {
-        public:
-        EventQueue();
-
-        void update();
-
-        const Event& front();
-
-        void pop();
-
-        bool empty();
-
-        EventQueueDelegate& getDelegate();
-
-        protected:
-            std::vector<Window*> windows;
-
-            EventQueueDelegate mDelegate;
-    };
-}
