+++ src/CrossWindow/UWP/UWPEnventQueue.cpp
@@ -1,0 +1,6 @@
+#include "UWPEventQueue.h"
+
+namespace xwin
+{
+	//Fill in with UWP event processing
+}
