+++ src/CrossWindow/UWP/UWPWindow.cpp
@@ -1,0 +1,14 @@
+#include "UWPWindow.h"
+
+namespace xwin
+{
+    bool Window::create(const WindowDesc& desc, EventQueue& queue)
+    {
+    // create stub
+    }
+
+    void Window::close()
+    {
+        
+    }
+}
